Refactor Gemini utilities into modular components with tests
